+++ CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Change `NodePath` internal representation to `PathBuf`
  - Remove unneeded '/' prefix strip in `StorePrefix`
  - **Breaking**: Remove storage trait implementations for `Arc<TStorage>`, now must explicitly deref
+ - Implement `Eq` and `PartialEq` for `DataType`
 
 ### Fixed
  - Fix store prefix to node path conversion and vice versa